trainer.py
"""
Trainer skeleton (simple PPO example using stable-baselines3)
This is a placeholder: replace DummyEnv with an AirSim-wrapped gym.Env for training.
"""
import gym
import numpy as np
from stable_baselines3 import PPO

class DummyEnv(gym.Env):
    def __init__(self):
        super().__init__()
        self.observation_space = gym.spaces.Box(low=-100, high=100, shape=(6,), dtype=np.float32)
        self.action_space = gym.spaces.Box(low=-1, high=1, shape=(3,), dtype=np.float32)

    def reset(self):
        return self.observation_space.sample()

    def step(self, action):
        obs = self.observation_space.sample()
        reward = 0.0
        done = False
        info = {}
        return obs, reward, done, info

def train_dummy():
    env = DummyEnv()
    model = PPO("MlpPolicy", env, verbose=1)
    model.learn(total_timesteps=10000)
    model.save("models/ppo_dummy")

if __name__ == '__main__':
    train_dummy()
