requirements.txt
numpy
matplotlib
pandas
pyzmq
ultralytics
opencv-python
stable-baselines3
torch
gymnasium
airsim==1.8.1
msgpack-rpc-python
